make ios compatible with new arch
