--- ios/LineLogin.h
@@ -1,4 +1,0 @@
-#ifndef LineLogin_h
-#define LineLogin_h
-
-#endif /* LineLogin_h */
