--- ios/LineLogin.m
@@ -1,33 +1,0 @@
-#import <Foundation/Foundation.h>
-#import "React/RCTBridgeModule.h"
-
-@interface RCT_EXTERN_MODULE(LineLogin, NSObject)
-RCT_EXTERN_METHOD(setup: (NSDictionary *)arguments
-                  resolver: (RCTPromiseResolveBlock)resolve
-                  rejecter: (RCTPromiseRejectBlock)reject
-                  )
-RCT_EXTERN_METHOD(login: (NSDictionary *)arguments
-                  resolver: (RCTPromiseResolveBlock)resolve
-                  rejecter: (RCTPromiseRejectBlock)reject
-                  )
-RCT_EXTERN_METHOD(logout: (RCTPromiseResolveBlock)resolve
-                  rejecter: (RCTPromiseRejectBlock)reject)
-RCT_EXTERN_METHOD(
-                  getFriendshipStatus: (RCTPromiseResolveBlock)resolve
-                  rejecter: (RCTPromiseRejectBlock)reject
-                  )
-RCT_EXTERN_METHOD(getCurrentAccessToken: (RCTPromiseResolveBlock)resolve
-                  rejecter: (RCTPromiseRejectBlock)reject)
-RCT_EXTERN_METHOD(
-                  getProfile: (RCTPromiseResolveBlock)resolve
-                  rejecter: (RCTPromiseRejectBlock)reject
-                  )
-RCT_EXTERN_METHOD(
-                  refreshAccessToken: (RCTPromiseResolveBlock)resolve
-                  rejecter: (RCTPromiseRejectBlock)reject
-                  )
-RCT_EXTERN_METHOD(
-                  verifyAccessToken: (RCTPromiseResolveBlock)resolve
-                  rejecter: (RCTPromiseRejectBlock)reject
-                  )
-@end
