+++ ios/LineLoginModule-Bridging-Header.h
@@ -1,0 +1,5 @@
+#import <React/RCTBridgeModule.h>
+
+#ifdef RCT_NEW_ARCH_ENABLED
+#import "RTNLineSpec.h"
+#endif
