--- ios/LineSetup.h
@@ -1,4 +1,0 @@
-#ifndef LineSetup_h
-#define LineSetup_h
-
-#endif /* LineSetup_h */
