--- ios/RNLine-Bridging-Header.h
@@ -1,7 +1,0 @@
-//
-//  Use this file to import your target's public headers that you would like to expose to Swift.
-//
-
-#import "React/RCTBridgeModule.h"
-#import <React/RCTBundleURLProvider.h>
-#import <React/RCTRootView.h>
